Leverage HttpMiddleware to customize content disposition rendering
