+++ web/src/main/scala/quasar/api/package.scala
@@ -25,6 +25,7 @@ import argonaut.{DecodeResult => _, _}, Argonaut._
 import org.http4s._
 import org.http4s.argonaut._
 import org.http4s.dsl.{Path => HPath, _}
+import org.http4s.headers.`Content-Disposition`
 import org.http4s.server._
 import org.http4s.server.staticcontent._
 import org.http4s.util._
@@ -126,6 +127,33 @@ package object api {
           err => BadRequest(Json("error" := "invalid request-headers: " + err)),
           service.run)
       }
+  }
+
+  // [#1861] Workaround to support a small slice of RFC 5987 for this isolated case
+  object RFC5987ContentDispositionRender extends HttpMiddleware {
+    // NonUnitStatements due to http4s's Writer
+    @SuppressWarnings(Array("org.wartremover.warts.NonUnitStatements"))
+    final case class ContentDisposition(dispositionType: String, parameters: Map[String, String])
+      extends Header.Parsed {
+      import org.http4s.util.Writer
+      override def key = `Content-Disposition`
+      override lazy val value = super.value
+      override def renderValue(writer: Writer): writer.type = {
+        writer.append(dispositionType)
+        parameters.foreach(p =>
+          p._1.endsWith("*").fold(
+            writer << "; " << p._1 << "=" << p._2,
+            writer << "; " << p._1 << "=\"" << p._2 << '"'))
+        writer
+      }
+    }
+
+    def apply(service: HttpService): HttpService =
+      service ∘ (resp => resp.headers.get(`Content-Disposition`).cata(
+        i => resp.copy(headers = resp.headers
+          .filter(_.name ≠ `Content-Disposition`.name)
+          .put(ContentDisposition(i.dispositionType, i.parameters))),
+        resp))
   }
 
   object Prefix {
